+++ src/test/mempool_tests.cpp
@@ -481,6 +481,7 @@ BOOST_AUTO_TEST_CASE(MempoolSizeLimitTest) {
     CTxMemPool pool;
     TestMemPoolEntryHelper entry;
     entry.dPriority = 10.0;
+    Amount feeIncrement = MEMPOOL_FULL_FEE_INCREMENT.GetFeePerK();
 
     CMutableTransaction tx1 = CMutableTransaction();
     tx1.vin.resize(1);
@@ -537,7 +538,7 @@ BOOST_AUTO_TEST_CASE(MempoolSizeLimitTest) {
                                CTransaction(tx3).GetTotalSize() +
                                    CTransaction(tx2).GetTotalSize());
     BOOST_CHECK_EQUAL(pool.GetMinFee(1).GetFeePerK(),
-                      maxFeeRateRemoved.GetFeePerK() + Amount(1000));
+                      maxFeeRateRemoved.GetFeePerK() + feeIncrement);
 
     CMutableTransaction tx4 = CMutableTransaction();
     tx4.vin.resize(2);
@@ -625,19 +626,19 @@ BOOST_AUTO_TEST_CASE(MempoolSizeLimitTest) {
     SetMockTime(42);
     SetMockTime(42 + CTxMemPool::ROLLING_FEE_HALFLIFE);
     BOOST_CHECK_EQUAL(pool.GetMinFee(1).GetFeePerK(),
-                      maxFeeRateRemoved.GetFeePerK() + Amount(1000));
+                      maxFeeRateRemoved.GetFeePerK() + feeIncrement);
     // ... we should keep the same min fee until we get a block
     pool.removeForBlock(vtx, 1);
     SetMockTime(42 + 2 * CTxMemPool::ROLLING_FEE_HALFLIFE);
     BOOST_CHECK_EQUAL(pool.GetMinFee(1).GetFeePerK(),
-                      (maxFeeRateRemoved.GetFeePerK() + Amount(1000)) / 2);
+                      (maxFeeRateRemoved.GetFeePerK() + feeIncrement) / 2);
     // ... then feerate should drop 1/2 each halflife
 
     SetMockTime(42 + 2 * CTxMemPool::ROLLING_FEE_HALFLIFE +
                 CTxMemPool::ROLLING_FEE_HALFLIFE / 2);
     BOOST_CHECK_EQUAL(
         pool.GetMinFee(pool.DynamicMemoryUsage() * 5 / 2).GetFeePerK(),
-        (maxFeeRateRemoved.GetFeePerK() + Amount(1000)) / 4);
+        (maxFeeRateRemoved.GetFeePerK() + feeIncrement) / 4);
     // ... with a 1/2 halflife when mempool is < 1/2 its target size
 
     SetMockTime(42 + 2 * CTxMemPool::ROLLING_FEE_HALFLIFE +
@@ -645,20 +646,8 @@ BOOST_AUTO_TEST_CASE(MempoolSizeLimitTest) {
                 CTxMemPool::ROLLING_FEE_HALFLIFE / 4);
     BOOST_CHECK_EQUAL(
         pool.GetMinFee(pool.DynamicMemoryUsage() * 9 / 2).GetFeePerK(),
-        (maxFeeRateRemoved.GetFeePerK() + Amount(1000)) / 8);
+        (maxFeeRateRemoved.GetFeePerK() + feeIncrement) / 8);
     // ... with a 1/4 halflife when mempool is < 1/4 its target size
-
-    SetMockTime(42 + 7 * CTxMemPool::ROLLING_FEE_HALFLIFE +
-                CTxMemPool::ROLLING_FEE_HALFLIFE / 2 +
-                CTxMemPool::ROLLING_FEE_HALFLIFE / 4);
-    BOOST_CHECK_EQUAL(pool.GetMinFee(1).GetFeePerK(), Amount(1000));
-    // ... but feerate should never drop below 1000
-
-    SetMockTime(42 + 8 * CTxMemPool::ROLLING_FEE_HALFLIFE +
-                CTxMemPool::ROLLING_FEE_HALFLIFE / 2 +
-                CTxMemPool::ROLLING_FEE_HALFLIFE / 4);
-    BOOST_CHECK_EQUAL(pool.GetMinFee(1).GetFeePerK(), Amount(0));
-    // ... unless it has gone all the way to 0 (after getting past 1000/2)
 
     SetMockTime(0);
 }