+++ doc/release-notes.md
@@ -8,3 +8,4 @@ This release includes the following features and fixes:
  - Remove support for `-sendfreetransactions`.
  - Remove unstable `estimatesmartfee` RPC.
  - Update berkley DB to 5.3 minimum. Developers should update their build environment accordingly.
+ - Remove `-incrementalrelayfee` option